// generated/nextjs/content-node/types.ts
export interface ContentNodeStorage {
  readonly get: (relation: string, key: string) => Promise<Record<string, unknown> | null>;
  readonly put: (relation: string, key: string, value: Record<string, unknown>) => Promise<void>;
  readonly delete: (relation: string, key: string) => Promise<boolean>;
  readonly find: (relation: string, filter?: Record<string, unknown>) => Promise<readonly Record<string, unknown>[]>;
}

export interface ContentNodeCreateInput {
  readonly node: string;
  readonly type: string;
  readonly content: string;
  readonly createdBy: string;
}

export interface ContentNodeCreateOutputOk {
  readonly variant: 'ok';
  readonly node: string;
}

export interface ContentNodeCreateOutputExists {
  readonly variant: 'exists';
  readonly message: string;
}

export type ContentNodeCreateOutput = ContentNodeCreateOutputOk | ContentNodeCreateOutputExists;

export interface ContentNodeUpdateInput {
  readonly node: string;
  readonly content: string;
}

export interface ContentNodeUpdateOutputOk {
  readonly variant: 'ok';
  readonly node: string;
}

export interface ContentNodeUpdateOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type ContentNodeUpdateOutput = ContentNodeUpdateOutputOk | ContentNodeUpdateOutputNotfound;

export interface ContentNodeDeleteInput {
  readonly node: string;
}

export interface ContentNodeDeleteOutputOk {
  readonly variant: 'ok';
  readonly node: string;
}

export interface ContentNodeDeleteOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type ContentNodeDeleteOutput = ContentNodeDeleteOutputOk | ContentNodeDeleteOutputNotfound;

export interface ContentNodeGetInput {
  readonly node: string;
}

export interface ContentNodeGetOutputOk {
  readonly variant: 'ok';
  readonly node: string;
  readonly type: string;
  readonly content: string;
  readonly metadata: string;
}

export interface ContentNodeGetOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type ContentNodeGetOutput = ContentNodeGetOutputOk | ContentNodeGetOutputNotfound;

export interface ContentNodeSetMetadataInput {
  readonly node: string;
  readonly metadata: string;
}

export interface ContentNodeSetMetadataOutputOk {
  readonly variant: 'ok';
  readonly node: string;
}

export interface ContentNodeSetMetadataOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type ContentNodeSetMetadataOutput = ContentNodeSetMetadataOutputOk | ContentNodeSetMetadataOutputNotfound;

export interface ContentNodeChangeTypeInput {
  readonly node: string;
  readonly type: string;
}

export interface ContentNodeChangeTypeOutputOk {
  readonly variant: 'ok';
  readonly node: string;
}

export interface ContentNodeChangeTypeOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type ContentNodeChangeTypeOutput = ContentNodeChangeTypeOutputOk | ContentNodeChangeTypeOutputNotfound;